Refactor validation exports
Intent: strengthen grammar routing coherence
Operators involved: [Mutation|Transition|Thol]
Affected invariants: [#4]
Key changes:
- Re-export grammar validation exceptions and results via tnfr.validation
- Redirect selectors, execution and structural modules to use validation imports
- Sync typing stubs with new validation surface
Expected risks/dissonances: minimal import regressions mitigated by targeted tests
Metrics: no direct C(t), Si, νf, phase impact expected
Equivalence map: tnfr.operators.grammar.* helpers -> tnfr.validation.*
