+++ src/tnfr/validation/__init__.py
@@ -39,6 +39,12 @@ class Validator(Protocol[SubjectT]):
 from .compatibility import CANON_COMPAT, CANON_FALLBACK
 from .grammar import (
     GrammarContext,
+    MutationPreconditionError,
+    RepeatWindowError,
+    SequenceValidationResult,
+    StructuralGrammarError,
+    TholClosureError,
+    TransitionCompatibilityError,
     apply_glyph_with_grammar,
     enforce_canonical_grammar,
     on_applied_glyph,
@@ -60,6 +66,12 @@ __all__ = (
     "Validator",
     "validate_sequence",
     "GrammarContext",
+    "StructuralGrammarError",
+    "RepeatWindowError",
+    "MutationPreconditionError",
+    "TholClosureError",
+    "TransitionCompatibilityError",
+    "SequenceValidationResult",
     "apply_glyph_with_grammar",
     "enforce_canonical_grammar",
     "on_applied_glyph",
