+++ src/tnfr/validation/__init__.pyi
@@ -5,6 +5,12 @@ from ..types import Glyph, TNFRGraph
 from .compatibility import CANON_COMPAT as CANON_COMPAT, CANON_FALLBACK as CANON_FALLBACK
 from .grammar import (
     GrammarContext,
+    MutationPreconditionError,
+    RepeatWindowError,
+    SequenceValidationResult,
+    StructuralGrammarError,
+    TholClosureError,
+    TransitionCompatibilityError,
     apply_glyph_with_grammar,
     enforce_canonical_grammar,
     on_applied_glyph,
@@ -38,6 +44,12 @@ __all__ = (
     "Validator",
     "validate_sequence",
     "GrammarContext",
+    "StructuralGrammarError",
+    "RepeatWindowError",
+    "MutationPreconditionError",
+    "TholClosureError",
+    "TransitionCompatibilityError",
+    "SequenceValidationResult",
     "apply_glyph_with_grammar",
     "enforce_canonical_grammar",
     "on_applied_glyph",
